refactor(about): remove legacy update checker file
